Initial Commit
So the idea of this is to create yet another FRC scouting app, partially
for my own knowledge and also to make an extension of something I tried
at the Sydney regional that seemed to be a decent stepping stone in
match prediction. Long story short I'm going to try and use a neural
network to calculate scouting weights and get match predictions, along
with having it as a general scouting summary program.